+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentOption+Images.swift
@@ -304,6 +304,8 @@ extension STPPaymentMethodType {
                 return .pm_type_touch_n_go
             case .trueMoney:
                 return .pm_type_truemoney
+            case .ngWallet:
+                return .pm_type_ng_wallet
             case .goPay:
                 return .pm_type_gopay
             case .grabPay: