+++ StripePayments/StripePayments/Source/PaymentHandler/STPPaymentHandler.swift
@@ -833,6 +833,7 @@ public class STPPaymentHandler: NSObject {
             .ngBankTransfer,
             .touchNGo,
             .trueMoney,
+            .ngWallet,
             .goPay,
             .gcash,
             .shopeePay,