+++ CHANGELOG.md
@@ -11,6 +11,7 @@ MINOR
 
 ### PaymentSheet
 * [Added] Added QRIS API bindings and support for payments in PaymentSheet.
+* [Added] Added Naira Wallet API bindings and support for payments in PaymentSheet.
 * [Added] Added ShopeePay API bindings and support for payments in PaymentSheet.
 * [Added] Added TrueMoney API bindings and support for payments and setup in PaymentSheet.
 * [Added] Added Touch 'n Go API bindings and support for payments and setup in PaymentSheet.